Moves variant to new folder
