+++ scout/server/blueprints/variant/templates/variant/variant.html
