+++ scout/server/blueprints/variants/controllers.py
@@ -31,7 +31,6 @@ LOG = logging.getLogger(__name__)
 class MissingVerificationRecipientError(Exception):
     pass
 
-
 def variants(store, institute_obj, case_obj, variants_query, page=1, per_page=50):
     """Pre-process list of variants."""
     variant_count = variants_query.count()
@@ -90,56 +89,9 @@ def str_variants(store, institute_obj, case_obj, variants_query, page=1, per_pag
     # Nothing unique to STRs on this level. Inheritance? yep, you will want it.
 
     # case bam_files for quick access to alignment view.
-    case_append_bam(store, case_obj)
+    case_append_bam(case_obj)
 
     return variants(store, institute_obj, case_obj, variants_query, page, per_page)
-
-def str_variant(store, institute_id, case_name, variant_id):
-    """Pre-process an STR variant entry for detail page.
-
-    Adds information to display variant
-
-    Args:
-        store(scout.adapter.MongoAdapter)
-        institute_id(str)
-        case_name(str)
-        variant_id(str)
-
-    Returns:
-        detailed_information(dict): {
-            'institute': <institute_obj>,
-            'case': <case_obj>,
-            'variant': <variant_obj>,
-            'overlapping_snvs': <overlapping_snvs>,
-            'manual_rank_options': MANUAL_RANK_OPTIONS,
-            'dismiss_variant_options': DISMISS_VARIANT_OPTIONS
-        }
-        """
-
-    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
-    variant_obj =  store.variant(variant_id)
-
-    # fill in information for pilup view
-    variant_case(store, case_obj, variant_obj)
-
-    variant_obj['callers'] = callers(variant_obj, category='str')
-
-    # variant_obj['str_ru']
-    # variant_obj['str_repid']
-    # variant_obj['str_ref']
-
-    variant_obj['comments'] = store.events(institute_obj, case=case_obj,
-                                           variant_id=variant_obj['variant_id'], comments=True)
-
-    return {
-        'institute': institute_obj,
-        'case': case_obj,
-        'variant': variant_obj,
-        'overlapping_snvs': overlapping_snvs,
-        'manual_rank_options': MANUAL_RANK_OPTIONS,
-        'dismiss_variant_options': DISMISS_VARIANT_OPTIONS
-    }
-
 
 def parse_variant(store, institute_obj, case_obj, variant_obj, update=False, genome_build='37',
                   get_compounds = True):
@@ -188,7 +140,7 @@ def parse_variant(store, institute_obj, case_obj, variant_obj, update=False, gen
 
     # We update the variant if some information was missing from loading
     # Or if symbold in reference genes have changed
-    if update and has_changed:
+    if (update and has_changed):
         variant_obj = store.update_variant(variant_obj)
 
     variant_obj['comments'] = store.events(institute_obj, case=case_obj,
@@ -357,155 +309,6 @@ def get_predictions(genes):
 
     return data
 
-def variant(store, institute_obj, case_obj, variant_id=None, variant_obj=None, add_case=True,
-            add_other=True, get_overlapping=True):
-    """Pre-process a single variant for the detailed variant view.
-
-    Adds information from case and institute that is not present on the variant
-    object
-
-    Args:
-        store(scout.adapter.MongoAdapter)
-        institute_obj(scout.models.Institute)
-        case_obj(scout.models.Case)
-        variant_id(str)
-        variant_obj(dict)
-        add_case(bool): If info about files case should be added
-        add_other(bool): If information about other causatives should be added
-        get_overlapping(bool): If overlapping svs should be collected
-
-    Returns:
-        variant_info(dict): {
-            'variant': <variant_obj>,
-            'causatives': <list(other_causatives)>,
-            'events': <list(events)>,
-            'overlapping_svs': <list(overlapping svs)>,
-            'manual_rank_options': MANUAL_RANK_OPTIONS,
-            'dismiss_variant_options': DISMISS_VARIANT_OPTIONS,
-            'ACMG_OPTIONS': ACMG_OPTIONS,
-            'evaluations': <list(evaluations)>,
-        }
-
-    """
-    # If the variant is already collected we skip this part
-    if not variant_obj:
-        default_panels = []
-        # Add default panel information to variant
-        for panel in case_obj['panels']:
-            if not panel.get('is_default'):
-                continue
-            panel_obj = store.gene_panel(panel['panel_name'], panel.get('version'))
-            if not panel:
-                LOG.warning("Panel {0} version {1} could not be found".format(
-                    panel['panel_name'], panel.get('version')))
-                continue
-            default_panels.append(panel_obj)
-
-        # NOTE this will query with variant_id == document_id, not the variant_id.
-        variant_obj = store.variant(variant_id, gene_panels=default_panels)
-
-    genome_build = case_obj.get('genome_build', '37')
-    if genome_build not in ['37','38']:
-        genome_build = '37'
-
-    if variant_obj is None:
-        return None
-    # Add information to case_obj
-    if add_case:
-        variant_case(store, case_obj, variant_obj)
-
-    # Collect all the events for the variant
-    events = list(store.events(institute_obj, case=case_obj, variant_id=variant_obj['variant_id']))
-    for event in events:
-        event['verb'] = VERBS_MAP[event['verb']]
-
-    other_causatives = set()
-    # Adds information about other causative variants
-    if add_other:
-        for other_causative in store.other_causatives(case_obj, variant_obj):
-            # avoid adding other causatives duplicates
-            other_causatives.add(tuple(other_causative.items()))
-        # convert set of tuples into list of dictionaries
-        other_causatives = [ dict(item) for item in list(other_causatives)]
-
-    variant_obj = parse_variant(store, institute_obj, case_obj, variant_obj, genome_build=genome_build)
-
-    variant_obj['end_position'] = end_position(variant_obj)
-    variant_obj['frequency'] = frequency(variant_obj)
-    variant_obj['clinsig_human'] = (clinsig_human(variant_obj) if variant_obj.get('clnsig')
-                                    else None)
-    variant_obj['thousandg_link'] = thousandg_link(variant_obj, genome_build)
-    variant_obj['exac_link'] = exac_link(variant_obj)
-    variant_obj['gnomad_link'] = gnomad_link(variant_obj)
-    variant_obj['swegen_link'] = swegen_link(variant_obj)
-    variant_obj['cosmic_link'] = cosmic_link(variant_obj)
-    variant_obj['beacon_link'] = beacon_link(variant_obj, genome_build)
-    variant_obj['ucsc_link'] = ucsc_link(variant_obj, genome_build)
-    variant_obj['alamut_link'] = alamut_link(variant_obj)
-    variant_obj['spidex_human'] = spidex_human(variant_obj)
-    variant_obj['expected_inheritance'] = expected_inheritance(variant_obj)
-    variant_obj['callers'] = callers(variant_obj, category='snv')
-
-    individuals = {individual['individual_id']: individual for individual in
-                   case_obj['individuals']}
-    for sample_obj in variant_obj['samples']:
-        individual = individuals[sample_obj.get('sample_id')]
-        if not individual:
-            return None
-        sample_obj['is_affected'] = True if individual['phenotype'] == 2 else False
-
-    gene_models = set()
-    variant_obj['disease_associated_transcripts'] = []
-
-    # Parse the gene models, both from panels and genes
-    for gene_obj in variant_obj.get('genes', []):
-        # Adds gene level links
-        parse_gene(gene_obj, genome_build)
-        omim_models = set()
-        for disease_term in gene_obj.get('disease_terms', []):
-            omim_models.update(disease_term.get('inheritance', []))
-        gene_obj['omim_inheritance'] = list(omim_models)
-
-        # Build strings for the disease associated transcripts from gene panel
-        for refseq_id in gene_obj.get('disease_associated_transcripts', []):
-            hgnc_symbol = (gene_obj['common']['hgnc_symbol'] if gene_obj.get('common') else
-                           gene_obj['hgnc_id'])
-            transcript_str = "{}:{}".format(hgnc_symbol, refseq_id)
-            variant_obj['disease_associated_transcripts'].append(transcript_str)
-
-        gene_models = gene_models | omim_models
-
-    if variant_obj.get('genetic_models'):
-        variant_models = set(model.split('_', 1)[0] for model in variant_obj['genetic_models'])
-        variant_obj['is_matching_inheritance'] = variant_models & gene_models
-
-    evaluations = []
-    for evaluation_obj in store.get_evaluations(variant_obj):
-        evaluation(store, evaluation_obj)
-        evaluations.append(evaluation_obj)
-
-    case_clinvars = store.case_to_clinVars(case_obj.get('display_name'))
-
-    if variant_id in case_clinvars:
-        variant_obj['clinvar_clinsig'] = case_clinvars.get(variant_id)['clinsig']
-
-    svs = []
-    if get_overlapping:
-        svs = (parse_variant(store, institute_obj, case_obj, variant_obj) for
-                            variant_obj in store.overlapping(variant_obj))
-
-    return {
-        'variant': variant_obj,
-        'causatives': other_causatives,
-        'events': events,
-        'overlapping_svs': svs,
-        'manual_rank_options': MANUAL_RANK_OPTIONS,
-        'dismiss_variant_options': DISMISS_VARIANT_OPTIONS,
-        'mosaic_variant_options': MOSAICISM_OPTIONS,
-        'ACMG_OPTIONS': ACMG_OPTIONS,
-        'evaluations': evaluations,
-    }
-
 def observations(store, loqusdb, case_obj, variant_obj):
     """Query observations for a variant."""
     composite_id = ("{this[chromosome]}_{this[position]}_{this[reference]}_"
@@ -579,29 +382,6 @@ def transcript_str(transcript_obj, gene_name=None):
     if gene_name:
         change_str = "{}:".format(gene_name) + change_str
     return change_str
-
-
-def end_position(variant_obj):
-    """Calculate end position for a variant."""
-    alt_bases = len(variant_obj['alternative'])
-    num_bases = max(len(variant_obj['reference']), alt_bases)
-    return variant_obj['position'] + (num_bases - 1)
-
-
-def frequency(variant_obj):
-    """Returns a judgement on the overall frequency of the variant.
-
-    Combines multiple metrics into a single call.
-    """
-    most_common_frequency = max(variant_obj.get('thousand_genomes_frequency') or 0,
-                                variant_obj.get('exac_frequency') or 0)
-    if most_common_frequency > .05:
-        return 'common'
-    elif most_common_frequency > .01:
-        return 'uncommon'
-    else:
-        return 'rare'
-
 
 def clinsig_human(variant_obj):
     """Convert to human readable version of CLINSIG evaluation."""
